+++ backend/data/src/main/kotlin/io/tolgee/dtos/request/BusinessEventReportRequest.kt
@@ -1,0 +1,16 @@
+package io.tolgee.dtos.request
+
+import javax.validation.constraints.NotBlank
+
+data class BusinessEventReportRequest(
+  @field:NotBlank
+  var eventName: String = "",
+
+  var anonymousUserId: String? = null,
+
+  var organizationId: Long? = null,
+
+  var projectId: Long? = null,
+
+  val data: Map<String, Any?>? = null
+)