+++ webapp/src/component/security/SignUp/SignUpView.tsx
@@ -20,7 +20,8 @@ import { container } from 'tsyringe';
 import { GlobalActions } from 'tg.store/global/GlobalActions';
 import { useRecaptcha } from './useRecaptcha';
 import { useApiMutation } from 'tg.service/http/useQueryApi';
-import { SplitContent, SPLIT_CONTENT_BREAK_POINT } from '../SplitContent';
+import { SPLIT_CONTENT_BREAK_POINT, SplitContent } from '../SplitContent';
+import { useReportOnce } from 'tg.hooks/useReportEvent';
 
 export type SignUpType = {
   name: string;
@@ -61,6 +62,8 @@ export const SignUpView: FunctionComponent = () => {
       disableBadRequestHandling: true,
     },
   });
+
+  useReportOnce('SIGN_UP_PAGE_OPENED');
 
   const onSubmit = async (data: SignUpType) => {
     const request = {