+++ webapp/src/views/organizations/members/OrganizationMembersView.tsx
@@ -1,5 +1,5 @@
-import { FunctionComponent, useState } from 'react';
-import { Box, Typography, Button } from '@mui/material';
+import { FunctionComponent, useEffect, useState } from 'react';
+import { Box, Button, Typography } from '@mui/material';
 import { T, useTranslate } from '@tolgee/react';
 
 import { PaginatedHateoasList } from 'tg.component/common/list/PaginatedHateoasList';
@@ -12,6 +12,7 @@ import { SimpleList } from 'tg.component/common/list/SimpleList';
 import { InviteDialog } from './InviteDialog';
 import { InvitationItem } from './InvitationItem';
 import { LINKS, PARAMS } from 'tg.constants/links';
+import { useReportEvent } from 'tg.hooks/useReportEvent';
 
 export const OrganizationMembersView: FunctionComponent = () => {
   const organization = useOrganization();
@@ -47,6 +48,12 @@ export const OrganizationMembersView: FunctionComponent = () => {
 
   const invitations =
     invitationsLoadable.data?._embedded?.organizationInvitations;
+
+  const reportEvent = useReportEvent();
+
+  useEffect(() => {
+    reportEvent('ORGANIZATION_MEMBERS_VIEW');
+  }, []);
 
   return (
     <BaseOrganizationSettingsView